requirements.txt
streamlit
pdfkit
jinja2
weasyprint ; platform_system == "Linux"
Pillow